+++ react-delegationdashboard/src/assets/styles/nodes/nodes.scss
