+++ react-delegationdashboard/src/assets/styles/stats/stats.scss
