+++ react-delegationdashboard/src/assets/styles/trim/trim.scss
