+++ react-delegationdashboard/src/components/NodesList/NodeTypes.ts
@@ -1,0 +1,10 @@
+export const nodeActions = {
+    unJail: { label: 'Unjail', transaction: 'unJail' },
+    unStake: { label: 'Unstake', transaction: 'unStake' },
+    reStake: { label: 'ReStake', transaction: 'reSreStakeUnStakedNodestake' },
+    unBond: { label: 'Unbond', transaction: 'unBond' }
+};
+
+export const inactiveNodeActions = {
+    stake: { label: 'Stake', transaction: 'stake' },
+};
